Removes task pkg Updating protobuf revisionId
