+++ scheduler/scheduler_test.go
@@ -1,4 +1,4 @@
-// +build legacy
+/// +build legacy
 
 /*
 http://www.apache.org/licenses/LICENSE-2.0.txt
